--- a/prometheus/cache/cache.go
+++ b/prometheus/cache/cache.go
@@ -43,61 +43,25 @@ var separatorByteSlice = []byte{model.SeparatorByte} // For convenient use with 
 // Use CachedTGatherer with classic Registry using NewMultiTRegistry and ToTransactionalGatherer helpers.
 // NOTE(bwplotka): Experimental, API and behaviour can change.
 type CachedTGatherer struct {
-	metricFamilyByName map[string]*family
+	metrics            map[uint64]*dto.Metric
+	metricFamilyByName map[string]*dto.MetricFamily
 	mMu                sync.RWMutex
 }
 
 func NewCachedTGatherer() *CachedTGatherer {
 	return &CachedTGatherer{
-		metricFamilyByName: map[string]*family{},
+		metrics:            make(map[uint64]*dto.Metric),
+		metricFamilyByName: map[string]*dto.MetricFamily{},
 	}
-}
-
-type family struct {
-	*dto.MetricFamily
-
-	metricsByHash map[uint64]*dto.Metric
-}
-
-// normalizeMetricFamilies returns a MetricFamily slice with empty
-// MetricFamilies pruned and the remaining MetricFamilies sorted by name within
-// the slice, with the contained Metrics sorted within each MetricFamily.
-func normalizeMetricFamilies(metricFamiliesByName map[string]*family) []*dto.MetricFamily {
-	for _, mf := range metricFamiliesByName {
-		if cap(mf.Metric) < len(mf.metricsByHash) {
-			mf.Metric = make([]*dto.Metric, 0, len(mf.metricsByHash))
-		}
-		mf.Metric = mf.Metric[:0]
-		for _, m := range mf.metricsByHash {
-			mf.Metric = append(mf.Metric, m)
-		}
-		sort.Sort(internal.MetricSorter(mf.Metric))
-	}
-
-	for _, mf := range metricFamiliesByName {
-		sort.Sort(internal.MetricSorter(mf.Metric))
-	}
-	names := make([]string, 0, len(metricFamiliesByName))
-	for name, mf := range metricFamiliesByName {
-		if len(mf.Metric) > 0 {
-			names = append(names, name)
-		}
-	}
-	sort.Strings(names)
-	result := make([]*dto.MetricFamily, 0, len(names))
-	for _, name := range names {
-		result = append(result, metricFamiliesByName[name].MetricFamily)
-	}
-	return result
 }
 
 // Gather implements TransactionalGatherer interface.
 func (c *CachedTGatherer) Gather() (_ []*dto.MetricFamily, done func(), err error) {
 	c.mMu.RLock()
 
-	// BenchmarkCachedTGatherer_Update shows, even for 1 million metrics among 1000 families
+	// BenchmarkCachedTGatherer_Update shows, even for 1 million metrics with 1000 families
 	// this is efficient enough (~300µs and ~50 kB per op), no need to cache it for now.
-	return normalizeMetricFamilies(c.metricFamilyByName), c.mMu.RUnlock, nil
+	return internal.NormalizeMetricFamilies(c.metricFamilyByName), c.mMu.RUnlock, nil
 }
 
 type Key struct {
@@ -159,9 +123,11 @@ func (c *CachedTGatherer) Update(reset bool, inserts []Insert, deletions []Key) 
 	c.mMu.Lock()
 	defer c.mMu.Unlock()
 
-	currMetricFamilyByName := c.metricFamilyByName
+	currMetrics := c.metrics
+	currMetricFamilies := c.metricFamilyByName
 	if reset {
-		currMetricFamilyByName = make(map[string]*family, len(c.metricFamilyByName))
+		currMetrics = make(map[uint64]*dto.Metric, len(c.metrics))
+		currMetricFamilies = make(map[string]*dto.MetricFamily, len(c.metricFamilyByName))
 	}
 
 	errs := prometheus.MultiError{}
@@ -173,35 +139,22 @@ func (c *CachedTGatherer) Update(reset bool, inserts []Insert, deletions []Key) 
 		}
 
 		// Update metric family.
-		mf, ok := currMetricFamilyByName[inserts[i].FQName]
-		oldMf, oldOk := c.metricFamilyByName[inserts[i].FQName]
+		mf, ok := c.metricFamilyByName[inserts[i].FQName]
 		if !ok {
-			if !oldOk {
-				mf = &family{
-					MetricFamily:  &dto.MetricFamily{},
-					metricsByHash: map[uint64]*dto.Metric{},
-				}
-				mf.Name = &inserts[i].FQName
-			} else if reset {
-				mf = &family{
-					MetricFamily:  oldMf.MetricFamily,
-					metricsByHash: make(map[uint64]*dto.Metric, len(oldMf.metricsByHash)),
-				}
-			}
+			mf = &dto.MetricFamily{}
+			mf.Name = &inserts[i].FQName
+		} else if reset {
+			// Reset metric slice, since we want to start from scratch.
+			mf.Metric = mf.Metric[:0]
 		}
-
 		mf.Type = inserts[i].ValueType.ToDTO()
 		mf.Help = &inserts[i].Help
 
-		currMetricFamilyByName[inserts[i].FQName] = mf
+		currMetricFamilies[inserts[i].FQName] = mf
 
 		// Update metric pointer.
 		hSum := inserts[i].hash()
-		m, ok := mf.metricsByHash[hSum]
-		if !ok && reset && oldOk {
-			m, ok = oldMf.metricsByHash[hSum]
-		}
-
+		m, ok := c.metrics[hSum]
 		if !ok {
 			m = &dto.Metric{Label: make([]*dto.LabelPair, 0, len(inserts[i].LabelNames))}
 			for j := range inserts[i].LabelNames {
@@ -249,7 +202,16 @@ func (c *CachedTGatherer) Update(reset bool, inserts []Insert, deletions []Key) 
 		if inserts[i].Timestamp != nil {
 			m.TimestampMs = proto.Int64(inserts[i].Timestamp.Unix()*1000 + int64(inserts[i].Timestamp.Nanosecond()/1000000))
 		}
-		mf.metricsByHash[hSum] = m
+		currMetrics[hSum] = m
+
+		if !reset && ok {
+			// If we did update without reset and we found metric in previous
+			// map, we know metric pointer exists in metric family map, so just continue.
+			continue
+		}
+
+		// Will be sorted later anyway, so just append.
+		mf.Metric = append(mf.Metric, m)
 	}
 
 	for _, del := range deletions {
@@ -258,18 +220,42 @@ func (c *CachedTGatherer) Update(reset bool, inserts []Insert, deletions []Key) 
 			continue
 		}
 
-		mf, ok := currMetricFamilyByName[del.FQName]
+		hSum := del.hash()
+		m, ok := currMetrics[hSum]
 		if !ok {
 			continue
 		}
+		delete(currMetrics, hSum)
 
-		hSum := del.hash()
-		if _, ok := mf.metricsByHash[hSum]; !ok {
+		mf, ok := currMetricFamilies[del.FQName]
+		if !ok {
+			// Impossible, but well...
+			errs.Append(fmt.Errorf("could not remove metric %s(%s) from metric family, metric family does not exists", del.FQName, del.LabelValues))
 			continue
 		}
-		delete(mf.metricsByHash, hSum)
+
+		toDel := -1
+		for i := range mf.Metric {
+			if mf.Metric[i] == m {
+				toDel = i
+				break
+			}
+		}
+
+		if toDel == -1 {
+			errs.Append(fmt.Errorf("could not remove metric %s(%s) from metric family, metric family does not have such metric", del.FQName, del.LabelValues))
+			continue
+		}
+
+		if len(mf.Metric) == 1 {
+			delete(currMetricFamilies, del.FQName)
+			continue
+		}
+
+		mf.Metric = append(mf.Metric[:toDel], mf.Metric[toDel+1:]...)
 	}
 
-	c.metricFamilyByName = currMetricFamilyByName
+	c.metrics = currMetrics
+	c.metricFamilyByName = currMetricFamilies
 	return errs.MaybeUnwrap()
 }