Lot's of extra structs here, so feedback welcome.

Things to do still here:

* [ ] Consider to add missed validation.