@bwplotka 2022-01-20
On the `rawCollector` I think I wanted to ensure that we don't put collector/gatherer which expects blocking registry into non blocking one, since we would have race on returned metricfamilies