@bwplotka 2022-01-20
Simplified API, added  tests