@bwplotka 2022-01-21
Found a race, changing a bit