@bwplotka 2022-01-21
Simplified the abstraction yet added possibility to use this cache for both event based and session based approach cc @dgrisonnet @beorn7 @kakkoyun 

PTAL (: