@bwplotka 2022-01-26
Ok, I think I finished optimizing this. I preferred memory efficiency over slight delay in updating and gathering. There is some room to improve it further, but it will add more complexity to the code.

Current results:

```
 benchstat -delta-test=none v6.txt
name                                                           time/op
CachedTGatherer_Update/Update_of_one_element_without_reset-12  13.1ms ± 0%
CachedTGatherer_Update/Update_of_all_elements_with_reset-12     309ms ± 0%
CachedTGatherer_Update/Gather-12                                422ms ± 0%

name                                                           alloc/op
CachedTGatherer_Update/Update_of_one_element_without_reset-12    208B ± 0%
CachedTGatherer_Update/Update_of_all_elements_with_reset-12    2.47kB ± 0%
CachedTGatherer_Update/Gather-12                               52.8kB ± 0%

name                                                           allocs/op
CachedTGatherer_Update/Update_of_one_element_without_reset-12    3.00 ± 0%
CachedTGatherer_Update/Update_of_all_elements_with_reset-12      0.00
CachedTGatherer_Update/Gather-12                                1.00k ± 0%
```

I think it's more than good for cadvisor/kubelet usage and potentially anyone else using this cache.