@bwplotka 2022-01-27
While those results are great I found some pain points in cadvisor usage of it. 

So apparently different kinds of metrics can be returned given the HTTP parameters. That makes it difficult for any caching logic which has only "memory" of the last state.

![image](https://user-images.githubusercontent.com/6950331/151341514-83a5c1a0-ba2d-40d9-b6c0-981abd1533de.png)

The solution we could do is to use two different caches (and live with allocating both) since we are lucky that only two options exist in cadvisor AFAIK. That is unfortunately not very sustainable, plus we keep mem for 2 caches. Probably better than it was with `const ` metrics on every scrape, but still not the best.

There is one thoughts if we want to improve this situation:

1. We could research a simple polling mechanism instead of cache. In the end, even in this cache we literally override all fields except labels. In this case, cadvisor might be able to just rebuild state on every needed interval while using pooling on e.g *Metric which if we keep polled items for longer, it can switch for the different request to different items. Something to think about.
 
All of this makes me think that such cache, for now, belongs to the user, not client_golang library. We should probably add `TransactionalGatherer` and other related improvements, but there are just many unknowns on API and behavior we need for this cache. I would propose moving cache for now to cadvisor repo and making this PR to only add those side `HandlerForTransactional` and `TransactionalGatherer` helpers, WDYT? 

cc @dgrisonnet @beorn7 @kakkoyun 

NOTE: I am 2 weeks sick off from today, so I might need to park this off. Anyone else is welcome to continue on this!