@bwplotka 2022-02-15
>  I think that the issue of having to maintain 2 caches is a problem for cadvisor to handle as it really depends on their use case. If there is really a lot of difference between the two versions of the metrics, then it makes sense to have two separate caches. But, as far as I can tell, this docker option allows filtering the metrics that are returned. So this could perhaps be improved by hooking into the Gather() method and filtering the results. Although that approach might come with a slight increase in CPU usage, it should be reasonable.

Makes sense 👍🏽 

>  That said, I think it is worth making the cache pluggable, by defining an interface based on how it is accessed by the Gatherers,

That's my point. This PR makes it pluggable in the sense that anyone can implement `TransactionalGatherer` with their own caching logic. My cache is just one implementation: https://github.com/prometheus/client_golang/blob/961f8ca5e65b8e7f890322fc13675ee5209a407d/prometheus/cache/cache.go#L52. Given it was done and tested only for cadvisor IMO we need more use cases to prove it deserves to be in this library.

> Taking that into account, an approach could be to use a queue to update the cache from the events

Sorry, I don't get this queue idea, not sure what's the benefits of it.[ I created API ](https://github.com/prometheus/client_golang/blob/961f8ca5e65b8e7f890322fc13675ee5209a407d/prometheus/cache/cache.go#L151) that have both collector and event-based approach (based on `reset` flag). Would that work? What is missing for even based approach?