@bwplotka 2022-02-23
Decided to close this for now. I will add another PR with just transactional interface.

Cache was efficient, but API causes extra work (those Insert structs) and errors like reusing same arrays. Let's try to come with something better in cadvisor first, only then port here.